Add command in root.
